// filmosearch.frontend/src/api/api.ts
export interface ActorDetailsVm {
    id?: number;
    firstNane?: string | undefined;
    lastName?: string | undefined;
    films?: FilmsLookupDto[] | undefined;
}

export interface ActorFilm {
    actorId?: number;
    actor?: Actor;
    filmId?: number;
    film?: Film;
}

export interface ActorListVm {
    actors?: ActorLookupDto[] | undefined;
}

export interface ActorLookupDto {
    id?: number;
    firstName?: string | undefined;
    lastName?: string | undefined;
}

export interface ActorsLookupDto {
    id?: number;
    firstName?: string | undefined;
    lastName?: string | undefined;
}

export interface AddFilmToActorDto {
    actorFirstName?: string | undefined;
    actorLastName?: string | undefined;
    filmTitle?: string | undefined;
}

export interface CreateActorDto {
    firstName?: string | undefined;
    lastName?: string | undefined;
}
// ...
export interface FilmsLookupDto {
    id?: number;
    title?: string | undefined;
}
// ...
export interface UpdateActorDto {
    id?: number;
    firstName?: string | undefined;
    lastName?: string | undefined;
}